% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{compute.header.palette.colors}
\alias{compute.header.palette.colors}
\title{Compute the colors for the CGA/EGA palette from the header.}
\usage{
compute.header.palette.colors(pcxheader, gfx_mode = "cga", raw_data = NULL)
}
\arguments{
\item{gfx_mode}{the graphics mode, one of 'cga', 'ega', or 'auto' to guess the mode from the data.}

\item{raw_data}{the image data, optional. Can be used to check the validity of the computed palette if given.}

\item{pcheader}{a PCX header instance}
}
\value{
integer matrix, the palette colors.
}
\description{
Compute the colors for the CGA/EGA palette from the header.
}
