% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_test_data}
\alias{download_test_data}
\title{Download test PCX files for the package.}
\usage{
download_test_data()
}
\value{
invisible list of files, usually ignored. This function is called for the side effect of downloading the data.
}
\description{
Download test PCX files for the package.
}
