% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{ega.palette.default16}
\alias{ega.palette.default16}
\title{Get default 16 colors 6-bit EGA palette.}
\usage{
ega.palette.default16()
}
\value{
named list with entries (1) `info`: data.frame with entries `index`, `color_names`, and `color_code_decimal` and (2) `colors`: 16x3 integer matrix representing RGB colors, in range 0-255.
}
\description{
Get default 16 colors 6-bit EGA palette.
}
\note{
EGA allowed one to create palettes by selecting 16 colors out of 64 possible ones. This is the default palette, see the \code{info$color_code_decimal} field to see which of the 64 colors were selected.
}
\keyword{internal}
