% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{guess.graphics.mode}
\alias{guess.graphics.mode}
\title{Guess graphics mode from header.}
\usage{
guess.graphics.mode(pcxheader)
}
\arguments{
\item{pcxheader}{PCX header instance}
}
\value{
one of 'CGA', 'EGA', or 'VGA'.
}
\description{
Guess graphics mode from header.
}
\keyword{internal}
