% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subbit.R
\name{ints.to.uint1}
\alias{ints.to.uint1}
\title{Split n 8 bit unsigned integers into 8n 1 bit unsigned integers.}
\usage{
ints.to.uint1(int8_vec)
}
\arguments{
\item{int8_vec}{integer vector of \code{n} values, all of which must be in range 0..255L.}
}
\value{
\code{8n} integers in range 0..1L.
}
\description{
Split n 8 bit unsigned integers into 8n 1 bit unsigned integers.
}
\keyword{internal}
