% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{ints.to.uint2}
\alias{ints.to.uint2}
\title{Split n 8 bit unsigned integers into 4n 2 bit unsigned integers.}
\usage{
ints.to.uint2(int8_vec)
}
\arguments{
\item{int8_vec}{integer vector of \code{n} values, all of which must be in range 0..255L.}
}
\value{
\code{4n} integers in range 0..3L.
}
\description{
Split n 8 bit unsigned integers into 4n 2 bit unsigned integers.
}
\keyword{internal}
