% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{is.pcx.indexed}
\alias{is.pcx.indexed}
\title{Determine from PCX header whether the image is indexed, i.e., whether it uses a fixed palette.}
\usage{
is.pcx.indexed(pcxheader)
}
\arguments{
\item{pcxheader}{a PCX header instance}
}
\value{
logical, whether the image is indexed.
}
\description{
Determine from PCX header whether the image is indexed, i.e., whether it uses a fixed palette.
}
\keyword{internal}
