% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{plot.pcx}
\alias{plot.pcx}
\title{S3 plot function for pcx image.}
\usage{
\method{plot}{pcx}(x, ...)
}
\arguments{
\item{x}{a pcx instance.}

\item{...}{extra args, not used.}
}
\description{
S3 plot function for pcx image.
}
