% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{print.pcx}
\alias{print.pcx}
\title{S3 print function for pcx image.}
\usage{
\method{print}{pcx}(x, ...)
}
\arguments{
\item{x}{a pcx instance.}

\item{...}{extra args, not used.}
}
\description{
S3 print function for pcx image.
}
