% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{print.pcxheader}
\alias{print.pcxheader}
\title{S3 print function for pcx header.}
\usage{
\method{print}{pcxheader}(x, ...)
}
\arguments{
\item{x}{a pcx.header instance.}

\item{...}{extra args, not used.}
}
\description{
S3 print function for pcx header.
}
