% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{read.pcx}
\alias{read.pcx}
\title{Read bitmap file in PCX format.}
\usage{
read.pcx(filepath, hdr = TRUE)
}
\arguments{
\item{filepath}{character string, path to the file including extension}

\item{hdr}{logical, whether to return full list with header}
}
\value{
array with color data, or pcx instance (named list) if `hdr` is `TRUE`
}
\description{
Read bitmap file in PCX format.
}
