% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_md5sum_file}
\alias{read_md5sum_file}
\title{Read md5sum file written by md5sum program.}
\usage{
read_md5sum_file(filepath)
}
\arguments{
\item{filepath}{character string, the path the the md5sum file}
}
\value{
data.frame with 3 columns: md5sum, is_binary, and filename.
}
\description{
Read md5sum file written by md5sum program.
}
\note{
File names entries must not include whitespace.
}
