% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subbit.R
\name{uint.subbits.as.uint}
\alias{uint.subbits.as.uint}
\title{Extract subbits from integer as unsigned int.}
\usage{
uint.subbits.as.uint(int8, start_index, stop_index)
}
\arguments{
\item{int8}{integer, must be in range 0L..255L. The integer is of course stored as a 32 bit integer internally in \code{R}, but it must contain a value between \code{0} and \code{255}, the other bits are ignored.}

\item{start_index}{integer in range 1L..32L, the start bit index.}

\item{stop_index}{integer in range 1L..32L, the stop bit index.}
}
\value{
an unsigned integer, the possible range depends on the number of bits.
}
\description{
Extract bits \code{start_index} to \code{stop_index} from an 8 bit integer.
}
\note{
Thanks to Julian_Hn at stackoverflow for this function.
}
\keyword{internal}
