% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pcx.R
\name{uint8split}
\alias{uint8split}
\title{Split each entry of 8 bit integer vector into 1, 2 or 4 bit integers.}
\usage{
uint8split(data_in, output_bits_per_int = 4L)
}
\arguments{
\item{data_in}{integer vector, containing values in range 0..255L (8 bit unsigned int values).}

\item{output_bits_per_int}{integer, one of 1L, 2L, 4L. The bits per output integer.}
}
\value{
vector of integer, the values range is smaller than the range of the 8 bit input values, and the number of values in greater.
}
\description{
Split each entry of 8 bit integer vector into 1, 2 or 4 bit integers.
}
\examples{
    uint8split(c(255, 8, 64, 13), 4L);
}
